+++ olutils/collection/deepdefaultdict.py
@@ -1,7 +1,8 @@
 """Extension of collections.defaultdict"""
-
 from collections import defaultdict as ddict
+from typing import Dict, Optional
 
+from olutils.typing import Factory
 from olutils.conversion.dictionary import basedict, dict2str
 
 
@@ -11,18 +12,16 @@ class DefaultDict(ddict):
     Overload collections.defaultdict conversion to base dictionary and pretty
     formatting.
     """
-    def to_dict(self):
+
+    def to_dict(self) -> Dict:
         """Return base dictionary from instance"""
         return basedict(self)
 
-    def pstring(self, *args, **kwargs):
+    def pstring(self, *args, **kwargs) -> str:
         """Return pretty string of instance
 
         Args:
             *args, **kwargs: @see `~olutils.dict2str`
-
-        Return:
-            (str)
         """
         return dict2str(self, *args, **kwargs)
 
@@ -35,12 +34,12 @@ class DefaultDict(ddict):
         print(self.pstring(*args, **kwargs))
 
 
-def defaultdict(default_factory, *args, **kwargs):
+def defaultdict(default_factory: Factory, *args, **kwargs) -> DefaultDict:
     """Return defaultdict given the default factory to produce new values"""
     return DefaultDict(default_factory, *args, **kwargs)
 
 
-def deepdefaultdict(default_factory, depth):
+def deepdefaultdict(default_factory: Factory, depth: int) -> Optional[DefaultDict]:
     """Return a multi ladder defaultdict given the leaf factory and depth"""
     if depth < 0:
         return None
@@ -48,4 +47,4 @@ def deepdefaultdict(default_factory, depth):
         return default_factory()
     if depth == 1:
         return defaultdict(default_factory)
-    return defaultdict(lambda: deepdefaultdict(default_factory, depth-1))
+    return defaultdict(lambda: deepdefaultdict(default_factory, depth - 1))