+++ olutils/collection/spestr.py
@@ -1,6 +1,0 @@
-"""Special strings3"""
-
-class SpeStr(str):
-    """String with no ' around repr"""
-    def __repr__(self):
-        return str(self)