+++ olutils/conversion/dictionary.py
@@ -1,42 +1,50 @@
 """Dictionary converters"""
+from typing import Any, Callable, Dict, Hashable
+
 from olutils.collection.functions import identity
 
 
-def basedict(__object, /, leafconv=identity):
+def basedict(__obj: Dict, /, leafconv: Callable = identity) -> Dict:
     """Return base dict from object (recursive)
 
     Args:
-        __object (dict)     : dict-like object to recursively convert
-        leafconv (callable) : function to convert leaves (not dict values)
-
-    Return:
-        (dict)
+        __obj (dict-like): dict-like object to recursively convert
+            for instance OrderedDict or defaultdict
+        leafconv: function to convert leaves (not dict values)
     """
-    if isinstance(__object, dict):
-        return {
-            key: basedict(value, leafconv=leafconv)
-            for key, value in __object.items()
-        }
-    return leafconv(__object)
+    return {
+        key: basedict(value, leafconv=leafconv)
+        if isinstance(value, dict)
+        else leafconv(value)
+        for key, value in __obj.items()
+    }
 
 
-def dict2str(__object, /, *, bullets="#*>-", indent="\t", prefix="",
-             keyconv=str, leafconv=str):
+def dict2str(
+    __obj: Any,
+    /,
+    *,
+    bullets: str = "#*>-",
+    indent: str = "\t",
+    prefix: str = "",
+    keyconv: Callable[[Hashable], str] = str,
+    leafconv: Callable[[Any], str] = str,
+) -> str:
     """Convert dict to pretty formatted string
 
     Args:
-        __object (dict) : dictionary to format to string
-        bullets (str)   : list of bullets to use for dict-ladders
-        indent (str)    : indent to add when going to next dict-ladder
-        prefix (str)    : prefix before each row in string
-        keyconv (func)  : dict-key string converter
-        leafconv (func) : dict-leaf string converter
+        __obj    : dictionary to format to string
+        bullets     : list of bullets to use for dict-ladders
+        indent      : indent to add when going to next dict-ladder
+        prefix      : prefix before each row in string
+        keyconv     : dict-key string converter
+        leafconv    : dict-leaf string converter
 
-    Return:
-        (str) pretty representation of __object
+    Returns:
+        (str) pretty representation of __obj
     """
-    if not isinstance(__object, dict):
-        return leafconv(__object)
+    if not isinstance(__obj, dict):
+        return leafconv(__obj)
 
     bullet = bullets[0] if bullets else ""
     exbullet = "" if bullet in ["", " "] else (bullet + " ")
@@ -44,10 +52,10 @@ def dict2str(__object, /, *, bullets="#*>-", indent="\t", prefix="",
 
     string = ""
     try:
-        key_maxsize = max(len(keyconv(key)) for key in __object.keys())
+        key_maxsize = max(len(keyconv(key)) for key in __obj.keys())
     except ValueError:
         return "<empty dict>"
-    for key, value in __object.items():
+    for key, value in __obj.items():
         if string:
             string += "\n"
         if isinstance(value, dict):
@@ -56,7 +64,7 @@ def dict2str(__object, /, *, bullets="#*>-", indent="\t", prefix="",
                 value,
                 bullets=nbullets,
                 indent=indent,
-                prefix=prefix+indent,
+                prefix=prefix + indent,
                 leafconv=leafconv,
                 keyconv=keyconv,
             )
@@ -66,8 +74,10 @@ def dict2str(__object, /, *, bullets="#*>-", indent="\t", prefix="",
                 string += " " + value_str
         else:
             string += (
-                prefix + exbullet
+                prefix
+                + exbullet
                 + keyconv(key).ljust(key_maxsize)
-                + ": " + leafconv(value)
+                + ": "
+                + leafconv(value)
             )
     return string