+++ olutils/conversion/error.py
@@ -1,7 +1,7 @@
 """Error converters"""
 
 
-def err2str(error, /, *, msg_only=False):
+def err2str(error: Exception, /, *, msg_only: bool = False) -> str:
     """Return string of error"""
     if msg_only:
         return str(error)