+++ olutils/conversion/temporality.py
@@ -1,9 +1,14 @@
-"""Temporality converters"""
-from collections.abc import Iterable
-from datetime import datetime, timedelta
-from dateutil.parser import parse
+"""Temporality converters
 
 # TODO : Handle timezones
+"""
+from collections.abc import Iterable as IterableABC
+from datetime import datetime, timedelta
+from typing import Iterable, Union
+
+from dateutil.parser import parse
+
+from olutils.typing import Number, TimeRepr
 
 DATE_REF = datetime(1970, 1, 1)
 
@@ -12,27 +17,29 @@ DAY = 24 * HOUR
 YEAR = 365.25 * DAY
 
 UNIT_TO_SEC = {
-    's': 1,
-    'sec': 1,
-    'second': 1,
-    'min': 60,
-    'minute': 60,
-    'h': HOUR,
-    'hour': HOUR,
-    'd': DAY,
-    'day': DAY,
-    'month': YEAR / 12,
-    'y': YEAR,
-    'year': YEAR,
+    "s": 1,
+    "sec": 1,
+    "second": 1,
+    "min": 60,
+    "minute": 60,
+    "h": HOUR,
+    "hour": HOUR,
+    "d": DAY,
+    "day": DAY,
+    "month": YEAR / 12,
+    "y": YEAR,
+    "year": YEAR,
 }
 
 
-def secs2unit(secs, /, unit):
+def secs2unit(
+    secs: Union[Number, Iterable[Number]], /, unit: str
+) -> Union[TimeRepr, Iterable[TimeRepr]]:
     """Convert number of seconds to given unit
 
     Args:
-        secs (int|float|iterable): number of seconds
-        unit (str): unit to convert to
+        secs: number of seconds
+        unit: unit to convert to
 
     Available units:
         s, sec, second      -> second
@@ -48,16 +55,15 @@ def secs2unit(secs, /, unit):
         (TypeError) : secs-type not handled
         (ValueError): unit not handled
 
-    Return:
-        (object)
-            if secs is int|float
-                if unit is dt: return datetime
-                elif unit is timedelta: return timedelta
-                else: return int|float
-            elif secs is iterable
-                if secs is list|set|tuple: return list|set|tuple
-                elif secs is ndarray & unit is not dt|td: return ndarray
-                else: return map-object
+    Returns:
+        if secs is number
+            if unit is dt: return datetime
+            elif unit is timedelta: return timedelta
+            else: return int|float
+        elif secs is iterable
+            if secs is list|set|tuple: return list|set|tuple
+            elif secs is ndarray & unit is not dt|td: return ndarray
+            else: return map-object
     """
     try:
         divisor = UNIT_TO_SEC[unit]
@@ -70,30 +76,30 @@ def secs2unit(secs, /, unit):
             return ts2dt(secs)
         if unit in ["td", "timedelta"]:
             return timedelta(seconds=float(secs))
-        raise ValueError("Unknown time unit '%s'" % unit)
+        raise ValueError(f"Unknown time unit '{unit}'")
     except TypeError:
         if isinstance(secs, (list, set, tuple)):
             return type(secs)(secs2unit(tick, unit) for tick in secs)
-        if not isinstance(secs, str) and isinstance(secs, Iterable):
+        if not isinstance(secs, str) and isinstance(secs, IterableABC):
             return map(lambda tick: secs2unit(tick, unit), secs)
         raise TypeError(f"Can't convert type {type(secs)}") from None
 
 
-def dt2ts(dt, /):
+def dt2ts(dt: datetime, /) -> float:
     """Return seconds since the Unix Epoch"""
-    return (dt-DATE_REF).total_seconds()
+    return (dt - DATE_REF).total_seconds()
 
 
-def ts2dt(ts, /):
+def ts2dt(ts: Number, /) -> datetime:
     """Return datetime from seconds since the Unix Epoch"""
     return DATE_REF + timedelta(seconds=float(ts))
 
 
-def str2dt(timestr, /, *args, **kwargs):
+def str2dt(timestr: str, /, *args, **kwargs) -> datetime:
     """Return datetime from string
 
     Args:
-        timestr (str)   : string describing a datetime
+        timestr: string describing a datetime
         *args, **kwargs : @see `dateutil.parser.parse`
     """
     return parse(timestr, *args, **kwargs)