+++ tests/test_collection/test_flatstr.py
@@ -1,0 +1,5 @@
+import olutils.collection as lib
+
+
+def test_FlatStr():
+    assert repr(lib.FlatStr("Hello")) == "Hello"