+++ tests/test_storing/test_csv.py
@@ -7,7 +7,7 @@ import olutils.storing as lib
 
 
 TMP_DIR = "tmp"
-MOCK_DIR = os.path.join("olutils", "tests", "mockups")
+MOCK_DIR = os.path.join("tests", "mockups")
 
 
 def assert_content_equal(path, content):